pyproject.toml
[tool.poetry]
name = "fast-poster"
version = "0.1.0"
description = ""
authors = ["seaung <seaung.c0m7de@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "9.0.1"
requests = "2.27.1"
requests-cache = "0.9.3"
qrcode = "7.3.1"
tornado = "6.1"
PyYAML = "6.0"
Flask = "^2.1.1"
Flask-SQLAlchemy = "^2.5.1"
PyMySQL = "^1.0.2"
DBUtils = "^3.0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
